requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
blinker==1.9.0
greenlet==3.2.4
click==8.3.1
typing_extensions==4.15.0